ReportOpenReal/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.5
djangorestframework==3.14.0
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
idna==3.4
mysqlclient==2.1.1
pika==1.3.1
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
